requirements.txt
# Database libraries
psycopg2-binary==2.9.9
SQLAlchemy==2.0.24
GeoAlchemy2==0.14.2

# Data processing libraries
pandas==2.1.3
numpy==1.26.2
geopandas==0.14.1
shapely==2.0.2

# API libraries
sodapy==2.2.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Flask for web app
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

# Redis for caching
redis==5.0.1

Flask==2.3.3
pandas==2.0.3
numpy==1.24.4
requests==2.31.0
python-dateutil==2.8.2
matplotlib==3.7.3
scipy==1.10.1
geopy==2.3.0
Werkzeug==2.3.7
gunicorn==21.2.0